Move `ManualOkOr` into `Methods` lint pass
